rejigging static queries so we don't load as much stuff in browser
